requirements.txt
selenium
pyserial
